rename test.cpp with recoder.cpp
