+++ apps/surfelwarp_kinectdk/CMakeLists.txt
@@ -5,7 +5,7 @@
 
 cuda_add_executable(surfelwarp_kinectdk main.cpp)
 
-cuda_add_executable(surfelwarp_kinectdk_test test.cpp)
+cuda_add_executable(surfelwarp_kinectdk_recorder recorder.cpp)
 
 #if(MSVC)
 #	set(CUDA_LIBRARY_DIR "${CUDA_TOOLKIT_ROOT_DIR}/lib/x64")
@@ -17,7 +17,7 @@ target_link_libraries(surfelwarp_kinectdk imgproc_lib)
 target_link_libraries(surfelwarp_kinectdk core_lib)
 target_link_libraries(surfelwarp_kinectdk visualization_lib)
 
-target_link_libraries(surfelwarp_kinectdk_test 
+target_link_libraries(surfelwarp_kinectdk_recorder 
     common_lib
     math_lib
     imgproc_lib
